Move generic network splitting functionality to network package
Functions to split network into smaller subnets is generic and needed in
several different resources (e.g. both Cluster and MachineDeployment IPAMs) so
it makes sense to have it in shared package.
