+++ service/network/splitter.go
@@ -1,0 +1,47 @@
+package network
+
+import (
+	"math/bits"
+	"net"
+
+	"github.com/giantswarm/ipam"
+	"github.com/giantswarm/microerror"
+)
+
+// CalculateSubnetMask calculates new subnet mask to accommodate n subnets.
+func CalculateSubnetMask(networkMask net.IPMask, n uint) (net.IPMask, error) {
+	if n == 0 {
+		return nil, microerror.Maskf(invalidParameterError, "divide by zero")
+	}
+
+	// Amount of bits needed to accommodate enough subnets for public and
+	// private subnet in each AZ.
+	subnetBitsNeeded := bits.Len(n - 1)
+
+	maskOnes, maskBits := networkMask.Size()
+	if subnetBitsNeeded > maskBits-maskOnes {
+		return nil, microerror.Maskf(invalidParameterError, "no room in network mask %s to accommodate %d subnets", networkMask.String(), n)
+	}
+
+	return net.CIDRMask(maskOnes+subnetBitsNeeded, maskBits), nil
+}
+
+// Split returns n subnets from network.
+func Split(network net.IPNet, n uint) ([]net.IPNet, error) {
+	mask, err := CalculateSubnetMask(network.Mask, n)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	var subnets []net.IPNet
+	for i := uint(0); i < n; i++ {
+		subnet, err := ipam.Free(network, mask, subnets)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+
+		subnets = append(subnets, subnet)
+	}
+
+	return subnets, nil
+}
